Moved back to plugin structure inside the package
