+++ ptypy/accelerate/cuda_pycuda/plugins/DM_pycuda_object_regul.py
