+++ templates/moonflower_test_dm_object_regul.py
@@ -8,12 +8,12 @@ from ptypy.core import Ptycho
 from ptypy import utils as u
 p = u.Param()
 
-import DM_object_regul 
-import DM_pycuda_object_regul
+from ptypy.plugins import DM_object_regul 
+from ptypy.accelerate.cuda_pycuda.plugins import DM_pycuda_object_regul
 import numpy as np
 
-#ny,nx = (512,512)
-ny,nx = (492,492)
+ny,nx = (512,512)
+#ny,nx = (492,492)
 xx,yy = np.meshgrid(np.arange(nx)-nx//2, np.arange(ny)-ny//2)
 mask = xx**2 + yy**2 > (150)**2
 mask = np.expand_dims(mask,0)
@@ -49,7 +49,7 @@ p.scans.MF.data.psf = 0.
 # attach a reconstrucion engine
 p.engines = u.Param()
 p.engines.engine00 = u.Param()
-p.engines.engine00.name = 'DM_object_regul'
+p.engines.engine00.name = 'DM_pycuda_object_regul'
 p.engines.engine00.numiter = 80
 p.engines.engine00.object_regul_mask = mask
 p.engines.engine00.object_regul_fill = 0.
